Correct api url, profile sorting, new bug reporter
The api was creating an extra '/' at the root of each request.

The sorting select on profile page had a typo, breaking a specific sort
select.

New Bug Report was not reading the user correctly on form submission.
It has been updated to read the name from the saved client info.